+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `approvalsService.closeFormApprovalFlowInstance()`
+
 ## [2.0.0] - 2022-08-28
 
 ### Removed